Refactor elb
