+++ src/modules/aws_elb@0.0.1/entity/aws_listener.ts
@@ -2,15 +2,20 @@ import {
   Column,
   Entity,
   JoinColumn,
-  JoinTable,
-  ManyToMany,
   ManyToOne,
   PrimaryGeneratedColumn,
 } from 'typeorm'
 
-import { AwsAction, } from './aws_action'
 import { AwsLoadBalancer, } from './aws_load_balancer'
-import { ProtocolEnum, } from './aws_target_group'
+import { AwsTargetGroup, ProtocolEnum, } from './aws_target_group'
+
+export enum ActionTypeEnum {
+  // AUTHENTICATE_COGNITO = "authenticate-cognito",
+  // AUTHENTICATE_OIDC = "authenticate-oidc",
+  // FIXED_RESPONSE = "fixed-response",
+  FORWARD = "forward",
+  // REDIRECT = "redirect"
+}
 
 @Entity()
 export class AwsListener {
@@ -35,9 +40,17 @@ export class AwsListener {
   })
   protocol: ProtocolEnum;
 
-  @ManyToMany(() => AwsAction, { cascade: true, })
-  @JoinTable()
-  defaultActions?: AwsAction[];
+  @Column({
+    type: 'enum',
+    enum: ActionTypeEnum,
+  })
+  actionType: ActionTypeEnum;
+
+  @ManyToOne(() => AwsTargetGroup)
+  @JoinColumn({
+    name: 'target_group_id',
+  })
+  targetGroup: AwsTargetGroup;
 
   // TODO: tbd
   // Certificates?: Certificate[];
