+++ src/modules/aws_elb@0.0.1/entity/index.ts
@@ -1,4 +1,3 @@
-export * from './aws_action'
 export * from './aws_listener'
 export * from './aws_load_balancer'
 export * from './aws_target_group'
